New transform pipeline
Remove legacy numpy operation. Read the images using torchvision instead of skimage. The images are loaded as tensors and all transformations are done as tensors. Previously we loaded the images as numpy arrays and we performed a numpy to tensor conversion. This version is around ~7% faster and uses less RAM due to performing less copies of the data.